Rename ebeye_urllib script
