+++ tools/ebi_tools/ebi_metagenomics_run_downloader.xml
@@ -9,7 +9,7 @@
 
     <command detect_errors="exit_code"><![CDATA[
         '$__tool_directory__/download_ebi_metagenomics_run_data'
-            '$__tool_directory__/ebeye_urllib3.py'
+            '$__tool_directory__/ebeye_urllib.py'
             '$run_id'
             '$information.type'
             '$information.to_download'
